fix bug of genMove in the root node

+++ src/genMove.hpp
@@ -35,15 +35,15 @@ public:
         sortType = 0;
     }
     step(int fromPos,int toPos,int fromPiece,int toPiece,int vl = 0){
-        this->fromPos = (uint8)fromPos;
-        this->toPos = (uint8)toPos;
-        this->fromPiece = (int8)fromPiece;
-        this->toPiece = (int8)toPiece;
+        this->fromPos = fromPos;
+        this->toPos = toPos;
+        this->fromPiece = fromPiece;
+        this->toPiece = toPiece;
         this->vl = vl;
         sortType = 0;
     }
     void printMove() const{
-        cout<<setw(3)<<(int)fromPiece<<" from "<<(int)fromPos<<" to "<<(int)toPos <<" and eat "<<(int)toPiece<<" with vl = "<<vl<<endl;
+        cout<<setw(3)<<fromPiece<<" from "<<fromPos<<" to "<<toPos <<" and eat "<<toPiece<<" with vl = "<<vl<<endl;
      }
     static void printMoveList(vector<step>& moveList){
         cout<<"--------------------------------------"<<endl;
@@ -65,10 +65,10 @@ public:
                 (toPiece == otherMove.toPiece);
     }
 public:
-    uint8 fromPos;
-    uint8 toPos;
-    int8 fromPiece;
-    int8 toPiece;
+    int fromPos;
+    int toPos;
+    int fromPiece;
+    int toPiece;
     int vl;
     int sortType;
 };
@@ -449,6 +449,7 @@ private:
                 const int targetList[4] = {leftTarget,rightTarget,upTarget,downTarget};
                 for(int a = 0; a < 4; a++){
                     const int mayEatToPos = p.bitBoard.getRayTargetPos(rookPos, targetList[a], 0);
+                    //p.bitBoard.printBitBoard();
                     if(mayEatToPos != -1){
                         const int toPiece = p.board.getPieceByPos(mayEatToPos);
                         if(rookPiece * toPiece < 0){
